+++ smarts/sstudio/types.py
@@ -515,7 +515,7 @@ class Trip:
     depart: float = 0
     """Start time in seconds."""
     actor: TrafficActor = field(init=False)
-    """The traffic actor for the vehicle"""
+    """The traffic actor model (usually vehicle) that will be used for the trip."""
 
     def __post_init__(self):
         object.__setattr__(